+++ frontend/openchat-agent/src/constants.ts
@@ -1,0 +1,6 @@
+export const MAX_EVENTS = 150;
+export const EVENT_PAGE_SIZE = 50;
+export const MAX_MISSING = 30;
+export const OPENCHAT_BOT_USER_ID = "zzyk3-openc-hatbo-tq7my-cai";
+export const OPENCHAT_BOT_USERNAME = "OpenChatBot";
+export const OPENCHAT_BOT_AVATAR_URL = "assets/robot.svg";
