+++ frontend/openchat-agent/src/domain/user/index.ts
@@ -1,1 +1,2 @@
 export * from "./user";
+export { extractUserIdsFromMentions, missingUserIds } from "./user.utils";
