+++ frontend/openchat-agent/src/domain/user/user.utils.spec.ts
@@ -1,27 +1,13 @@
 import type { PartialUserSummary, UserLookup } from "./user";
-import { init, addMessages, _ } from "svelte-i18n";
 import {
-    buildUsernameList,
     compareUsername,
     extractUserIdsFromMentions,
     missingUserIds,
     userIsOnline,
 } from "./user.utils";
-import { get } from "svelte/store";
-
-addMessages("en", {
-    you: "you",
-    andNMore: ", and {n} more",
-    reactions: {
-        andYou: ", and you",
-    },
-});
 
 const now = Date.now();
 jest.setSystemTime(now);
-init({
-    fallbackLocale: "en",
-});
 
 const lookup: UserLookup = {
     a: {
@@ -113,41 +99,5 @@ describe("compare username", () => {
             undefined,
             undefined,
         ]);
-    });
-});
-
-describe("build username list", () => {
-    test("and you and more", () => {
-        const userIds = Object.entries(lookup).map(([k, _]) => k);
-        userIds.push("beta");
-
-        const result = buildUsernameList(get(_), new Set(userIds), "alpha", lookup, 2);
-
-        expect(result).toEqual("a, b, and you, and 2 more");
-    });
-
-    test("show all", () => {
-        const userIds = Object.entries(lookup).map(([k, _]) => k);
-
-        const result = buildUsernameList(get(_), new Set(userIds), undefined, lookup);
-
-        expect(result).toEqual("a, b, julian_jelfs, alpha");
-    });
-
-    test("don't show 1 more", () => {
-        const userIds = Object.entries(lookup).map(([k, _]) => k);
-
-        const result = buildUsernameList(get(_), new Set(userIds), undefined, lookup, 3);
-
-        expect(result).toEqual("a, b, julian_jelfs, alpha");
-    });
-
-    test("do show 1 more if missing", () => {
-        const userIds = Object.entries(lookup).map(([k, _]) => k);
-        userIds.push("beta");
-
-        const result = buildUsernameList(get(_), new Set(userIds), undefined, lookup);
-
-        expect(result).toEqual("a, b, julian_jelfs, alpha, and 1 more");
     });
 });
