+++ frontend/openchat-agent/src/index.ts
@@ -1,4 +1,12 @@
 export * from "./domain";
 export * from "./services";
 
-export { buildCryptoTransferText, buildTransactionUrl } from "./domain/chat/chat.utils";
+export {
+    compareChats,
+    getContentAsText,
+    getFirstUnreadMessageIndex,
+    updateArgsFromChats,
+    emptyChatMetrics,
+    indexRangeForChat,
+    userIdsFromEvents,
+} from "./utils/chat";
