+++ frontend/openchat-agent/src/services/group/group.client.ts
@@ -71,8 +71,8 @@ import { cachingLocallyDisabled, Database } from "../../utils/caching";
 import { Principal } from "@dfinity/principal";
 import { apiMessageContent, apiOptional, apiUser } from "../common/chatMappers";
 import { DataClient } from "../data/data.client";
-import { identity, mergeGroupChatDetails } from "../../domain/chat/chat.utils";
-import { MAX_EVENTS } from "../../domain/chat/chat.utils.shared";
+import { identity, mergeGroupChatDetails } from "../../utils/chat";
+import { MAX_EVENTS } from "../../constants";
 import type { SearchGroupChatResponse } from "../../domain/search/search";
 import { getChatEventsInLoop } from "../common/chatEvents";
 import { profile } from "../common/profiling";
