+++ frontend/openchat-agent/src/services/ledger/ledger.client.ts
@@ -4,18 +4,18 @@ import { CandidService } from "../candidService";
 import type { ILedgerClient } from "./ledger.client.interface";
 import type { Tokens } from "../../domain/crypto";
 import { hexStringToBytes, identity } from "../../utils/mapping";
-import type { OpenChatConfig } from "../../config";
+import type { AgentConfig } from "../../config";
 
 export class LedgerClient extends CandidService implements ILedgerClient {
     private service: LedgerService;
 
-    private constructor(identity: Identity, config: OpenChatConfig, canisterId: string) {
+    private constructor(identity: Identity, config: AgentConfig, canisterId: string) {
         super(identity);
 
         this.service = this.createServiceClient<LedgerService>(idlFactory, canisterId, config);
     }
 
-    static create(identity: Identity, config: OpenChatConfig, canisterId: string): ILedgerClient {
+    static create(identity: Identity, config: AgentConfig, canisterId: string): ILedgerClient {
         return new LedgerClient(identity, config, canisterId);
     }
