+++ frontend/openchat-agent/src/services/snsGovernance/sns.governance.client.ts
@@ -4,12 +4,12 @@ import { CandidService } from "../candidService";
 import type { ISnsGovernanceClient } from "./sns.governance.client.interface";
 import type { ListNervousSystemFunctionsResponse } from "../../domain/chat/chat";
 import { nervousSystemFunctions } from "../common/chatMappers";
-import type { OpenChatConfig } from "../../config";
+import type { AgentConfig } from "../../config";
 
 export class SnsGovernanceClient extends CandidService implements ISnsGovernanceClient {
     private service: SnsGovernanceService;
 
-    private constructor(identity: Identity, config: OpenChatConfig, canisterId: string) {
+    private constructor(identity: Identity, config: AgentConfig, canisterId: string) {
         super(identity);
 
         this.service = this.createServiceClient<SnsGovernanceService>(
@@ -21,7 +21,7 @@ export class SnsGovernanceClient extends CandidService implements ISnsGovernance
 
     static create(
         identity: Identity,
-        config: OpenChatConfig,
+        config: AgentConfig,
         canisterId: string
     ): ISnsGovernanceClient {
         return new SnsGovernanceClient(identity, config, canisterId);
