+++ frontend/openchat-agent/src/utils/base64.ts
