+++ frontend/openchat-agent/src/utils/i18n.ts
@@ -1,0 +1,10 @@
+type InterpolationValues =
+    | Record<string, string | number | boolean | Date | null | undefined>
+    | undefined;
+interface MessageObject {
+    locale?: string;
+    format?: string;
+    default?: string;
+    values?: InterpolationValues;
+}
+export type MessageFormatter = (id: string, options?: MessageObject) => string;
