+++ frontend/openchat-client/src/index.ts
@@ -1,8 +1,7 @@
-export * from "./domain";
+export * from "openchat-agent";
 export * from "./stores";
 export { OpenChat } from "./openchat";
 export { OpenChatConfig } from "./config";
-export * from "./services";
 export * from "./events";
 
 export { ONE_GB } from "./stores/storage";
@@ -10,10 +9,6 @@ export { FilteredProposals } from "./stores/filteredProposals";
 export { MessageReadState } from "./stores/markRead";
 export { TypersByKey } from "./stores/typing";
 export { immutableStore } from "./stores/immutable";
-export { UnsupportedValueError } from "./utils/error";
 export { Dimensions, MAX_AUDIO_SIZE } from "./utils/media";
 export { MessageFormatter } from "./utils/i18n";
-export { Logger } from "./utils/logging";
-
-export { buildCryptoTransferText, buildTransactionUrl } from "./domain/chat/chat.utils";
 export { createMapStore } from "./stores/mapStore";
