+++ frontend/openchat-client/src/stores/filteredProposals.ts
@@ -1,4 +1,4 @@
-import type { ChatSummary, Proposal } from "../domain/chat/chat";
+import type { ChatSummary, Proposal } from "openchat-agent";
 import { writable } from "svelte/store";
 
 const storageKeyPrefix = "proposal_filters_";
