+++ frontend/openchat-client/src/stores/notifications.ts
@@ -1,4 +1,4 @@
-import type { NotificationStatus } from "src/domain";
+import type { NotificationStatus } from "openchat-agent";
 import { derived, writable } from "svelte/store";
 import { getSoftDisabled, storeSoftDisabled } from "../utils/caching";
