+++ frontend/openchat-client/src/stores/unconfirmed.ts
@@ -1,6 +1,6 @@
 import { get, writable } from "svelte/store";
 import { createSetStore } from "./setStore";
-import type { EventWrapper, Message } from "../domain/chat/chat";
+import type { EventWrapper, Message } from "openchat-agent";
 import { revokeObjectUrls } from "../domain/chat/chat.utils";
 
 export type UnconfirmedMessages = Record<
@@ -49,15 +49,13 @@ function createUnconfirmedStore() {
                         return {
                             ...state,
                             [key]: {
-                                messages: chatEvents.messages.filter(
-                                    (e) => {
-                                        if (e.event.messageId === messageId) {
-                                            revokeObjectUrls(e);
-                                            return false;
-                                        }
-                                        return true;
+                                messages: chatEvents.messages.filter((e) => {
+                                    if (e.event.messageId === messageId) {
+                                        revokeObjectUrls(e);
+                                        return false;
                                     }
-                                ),
+                                    return true;
+                                }),
                                 messageIds,
                             },
                         };
